+++ PlayerScore.cs
@@ -3,11 +3,11 @@ namespace ConfidencePoolAnalyzer
     class PlayerScore
     {
         public string Name;
-        public double Score;
+        public int Score;
         public int Rank;
         public double WeightedRank;
 
-        public PlayerScore(string name, double score)
+        public PlayerScore(string name, int score)
         {
             Name = name;
             Score = score;
